requirements.txt
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cycler==0.11.0
Flask==2.2.5
fonttools==4.38.0
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.5.3
mock==2.0.0
nltk==3.8.1
numpy==1.21.6
packaging==23.1
pbr==5.11.1
Pillow==9.5.0
pycocotools==2.0.6
pyparsing==3.0.9
python-dateutil==2.8.2
regex==2023.5.5
requests==2.13.0
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
threadpoolctl==3.1.0
torch==1.13.1
torchvision==0.14.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==2.0.2
Vocabulary==1.0.4
Werkzeug==2.2.3
zipp==3.15.0
